tests/PageIndexPyQt5/requirements.txt
PyQt5==5.15.10
openai>=1.0.0
pymupdf
PyPDF2
python-dotenv
tiktoken
pyyaml
python-docx
